Move thunks into thunks file from actions
